+++ packages/orq-rc/RELEASES.md
@@ -7909,3 +7909,13 @@ Based on:
 - [typescript v4.7.0-rc.34] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.34] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.34 - packages/orq-rc
+
+## 2026-03-31 12:19:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.35] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.35 - packages/orq-rc